Refactor core layer testing
