+++ test/config/ml-metadata/conn_config.pb
@@ -1,0 +1,6 @@
+connection_config {
+  sqlite {
+    filename_uri: '/tmp/shared/metadata.sqlite.db'
+    connection_mode: READWRITE_OPENCREATE
+  }
+}
